requirements.txt
# Discord bot
discord.py[voice]==2.6.4
PyNaCl==1.5.0

# AI
google-genai==1.64.0

# Utilities
python-dotenv==1.2.1
